Farms CRUD fully funtional. Fixed CreateModel for User

+++ Entities/Farm.cs
@@ -10,8 +10,9 @@ namespace WebApi.Entities
         public string Country { get; set;}
         public string Province { get; set;}
         public string City { get; set;}
-        public int Latitude { get; set;}
-        public int Longitude { get; set;}
+        public double Latitude { get; set;}
+        public double Longitude { get; set;}
         public bool isDeleted { get; set;}
+        
     }
 }
